text_to_wchar now private in kernel32\process.py
