+++ pywincffi/wintypes/functions.py
@@ -116,24 +116,3 @@ def socket_from_object(sock):
         sock = SOCKET()
         sock._cdata[0] = ffi.cast("SOCKET", fileno)
         return sock
-
-
-def text_to_wchar(text):
-    """
-    Converts ``text`` to ``wchar_t[len(text)]``.
-
-    :param text:
-        The text convert to ``wchar_t``.
-
-    :raises InputError:
-        Raised if the type of ``text`` is not a text type.  For
-        Python 2 this function expects unicode and for Python 3
-        it expects a string.
-    """
-    if not isinstance(text, text_type):
-        raise InputError(
-            "text", text,
-            message="Expected %r for `text`" % text_type)
-
-    ffi, _ = dist.load()
-    return ffi.new("wchar_t[%d]" % len(text), text)
